Refactoring the code and removing unwanted packages
