+++ cmd/k8s-bigip-ctlr/main_test.go
@@ -37,7 +37,7 @@ var _ = Describe("Main Tests", func() {
 				"--cis-config-cr=default/testcr",
 				"--log-level=INFO",
 				"--disable-teems=true",
-				"--insecure=true",
+				"--no-verify-ssl=true",
 				"--trusted-certs-cfgmap=default/foomap",
 				"--kubeconfig=/tmp/kubeconfig",
 				"--credentials-directory=/tmp/k8s-test-creds",
