--- pkg/resource/resourcecfg_suite_test.go
@@ -1,13 +1,0 @@
-package resource_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestResource(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Resource Configs Suite")
-}
