--- pkg/vxlan/vxlan_suite_test.go
@@ -1,13 +1,0 @@
-package vxlan_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestVxlan(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Vxlan Suite")
-}
