--- pkg/writer/writer_suite_test.go
@@ -1,13 +1,0 @@
-package writer_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestWriter(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Writer Suite")
-}
